[58417] feat: integrate pug build into vite as plugin
