+++ src/main/webapp/admin/v9/package.json
@@ -7,9 +7,9 @@
     },
     "scripts": {
         "clean": "rm -rf dist",
-        "dev": "npm run clean && vite build --mode development && node build-templates.mjs",
-        "prod": "npm run clean && vite build && node build-templates.mjs",
-        "watch": "npm run clean && (trap 'kill 0' INT TERM; vite build --mode development --watch & node build-templates.mjs --watch & wait)",
+        "dev": "npm run clean && vite build --mode development",
+        "prod": "npm run clean && vite build",
+        "watch": "npm run clean && vite build --mode development --watch",
         "postinstall": "./postinstall.sh"
     },
     "keywords": [],
