+++ src/main/webapp/admin/v9/src/scss/ninja.scss
@@ -220,3 +220,6 @@ $ai-color-hover: #E3DBFF;
 // HOTFIX
 @import "hotfix";
 
+.vite-test {
+    color: red;
+}
