+++ src/main/webapp/admin/v9/views/pages/dashboard/overview.pug
@@ -6,6 +6,8 @@ block content
         .ly-content-wrapper { padding-top: 34px; }
         .ly-content { height: calc(100vh - 34px); }
 
+    p.vite-test hahaha
+
     <div id="webjet-overview-dashboard">
         <a name="overview-dashboard-a" style="scroll-margin-top: 10em;"></a>
         <webjet-overview-dashboard th:v-bind:overviewUser="${overviewUser}" th:v-bind:overviewbackdata="${overviewBackData}" th:v-bind:overviewadmins="${overviewAdmins}" th:v-bind:overviewrecentpages="${overviewRecentPages}" th:v-bind:overviewchangedpages="${overviewChangedPages}" th:v-bind:overviewadminlog="${overviewAdminlog}" th:v-bind:overviewtodo="${overviewTodo}" th:v-bind:overviewcurrentsessions="${overviewCurrentSessions}"></webjet-overview-dashboard>
